move electromagnetics and waves to traits
