+++ src/geonum_mod.rs
@@ -7,19 +7,6 @@ use std::f64::consts::{PI, TAU};
 pub const TWO_PI: f64 = TAU;
 pub const EPSILON: f64 = 1e-10;
 
-// physical constants
-/// speed of light in vacuum (m/s)
-pub const SPEED_OF_LIGHT: f64 = 3.0e8;
-
-/// vacuum permeability (H/m)
-pub const VACUUM_PERMEABILITY: f64 = 4.0 * PI * 1e-7;
-
-/// vacuum permittivity (F/m)
-pub const VACUUM_PERMITTIVITY: f64 = 1.0 / (VACUUM_PERMEABILITY * SPEED_OF_LIGHT * SPEED_OF_LIGHT);
-
-/// vacuum impedance (Ω)
-pub const VACUUM_IMPEDANCE: f64 = VACUUM_PERMEABILITY * SPEED_OF_LIGHT;
-
 /// `Geonum` represents a single directed quantity in a specific blade direction:
 /// - `length`: the magnitude (can encode fractional participation)
 /// - `angle`: the orientation (in radians, mod 2π)
@@ -312,150 +299,6 @@ impl Geonum {
         (x, y)
     }
 
-    /// creates a field with 1/r^n falloff from a source
-    ///
-    /// useful for various physical fields that follow inverse power laws
-    ///
-    /// # args
-    /// * `charge` - source strength (charge, mass, etc.)
-    /// * `distance` - distance from source
-    /// * `power` - exponent for distance (1 for 1/r, 2 for 1/r², etc.)
-    /// * `angle` - field direction
-    /// * `constant` - physical constant multiplier
-    ///
-    /// # returns
-    /// a new geometric number representing the field
-    pub fn inverse_field(
-        charge: f64,
-        distance: f64,
-        power: f64,
-        angle: f64,
-        constant: f64,
-    ) -> Self {
-        let magnitude = constant * charge.abs() / distance.powf(power);
-        // Normalize angle calculation for negative charges
-        let direction = if charge >= 0.0 {
-            angle
-        } else {
-            // When angle is PI and we add PI, normalize to 0.0 rather than 2π
-            if angle == PI {
-                0.0
-            } else {
-                angle + PI
-            }
-        };
-
-        Self {
-            length: magnitude,
-            angle: direction,
-            blade: 1, // default to vector grade for fields
-        }
-    }
-
-    /// calculates electric potential at a distance from a point charge
-    ///
-    /// uses coulombs law for potential: V = k*q/r
-    /// where k = 1/(4πε₀)
-    ///
-    /// # args
-    /// * `charge` - electric charge in coulombs
-    /// * `distance` - distance from charge in meters
-    ///
-    /// # returns
-    /// the scalar potential (voltage) at the specified distance
-    pub fn electric_potential(charge: f64, distance: f64) -> f64 {
-        // coulomb constant k = 1/(4πε₀)
-        let k = 1.0 / (4.0 * PI * VACUUM_PERMITTIVITY);
-        k * charge / distance
-    }
-
-    /// calculates electric field at a distance from a point charge
-    ///
-    /// uses coulombs law: E = k*q/r²
-    /// where k = 1/(4πε₀)
-    ///
-    /// # args
-    /// * `charge` - electric charge in coulombs
-    /// * `distance` - distance from charge in meters
-    ///
-    /// # returns
-    /// a geometric number representing the electric field with:
-    /// * `length` - field magnitude
-    /// * `angle` - field direction (PI points radially outward for positive charge)
-    pub fn electric_field(charge: f64, distance: f64) -> Self {
-        let k = 1.0 / (4.0 * PI * VACUUM_PERMITTIVITY);
-        Self::inverse_field(charge, distance, 2.0, PI, k)
-    }
-
-    /// calculates the poynting vector using wedge product
-    ///
-    /// the poynting vector represents electromagnetic energy flux
-    /// traditionally calculated as S = E × B/μ₀ (cross product)
-    /// in geometric algebra, this is elegantly expressed with the wedge product
-    ///
-    /// # args
-    /// * `self` - the electric field as a geometric number
-    /// * `b_field` - the magnetic field as a geometric number
-    ///
-    /// # returns
-    /// a geometric number representing the poynting vector (energy flux)
-    pub fn poynting_vector(&self, b_field: &Self) -> Self {
-        // wedge product handles the cross product geometry in ga
-        let poynting = self.wedge(b_field);
-        Self {
-            length: poynting.length / VACUUM_PERMEABILITY,
-            angle: poynting.angle,
-            blade: poynting.blade,
-        }
-    }
-
-    /// creates a magnetic vector potential for a current-carrying wire
-    ///
-    /// # args
-    /// * `r` - distance from the wire
-    /// * `current` - current in the wire
-    /// * `permeability` - magnetic permeability of the medium
-    ///
-    /// # returns
-    /// a geometric number representing the magnetic vector potential
-    pub fn wire_vector_potential(r: f64, current: f64, permeability: f64) -> Self {
-        // A = (μ₀I/2π) * ln(r) in theta direction around wire
-        let magnitude = permeability * current * (r.ln()) / (2.0 * PI);
-        Self::from_polar(magnitude, PI / 2.0)
-    }
-
-    /// creates a magnetic field for a current-carrying wire
-    ///
-    /// # args
-    /// * `r` - distance from the wire
-    /// * `current` - current in the wire
-    /// * `permeability` - magnetic permeability of the medium
-    ///
-    /// # returns
-    /// a geometric number representing the magnetic field
-    pub fn wire_magnetic_field(r: f64, current: f64, permeability: f64) -> Self {
-        // B = μ₀I/(2πr) in phi direction circling the wire
-        let magnitude = permeability * current / (2.0 * PI * r);
-        Self::from_polar(magnitude, 0.0)
-    }
-
-    /// creates a scalar potential for a spherical electromagnetic wave
-    ///
-    /// # args
-    /// * `r` - distance from source
-    /// * `t` - time
-    /// * `wavenumber` - spatial frequency (k)
-    /// * `speed` - wave propagation speed
-    ///
-    /// # returns
-    /// a geometric number representing the scalar potential
-    pub fn spherical_wave_potential(r: f64, t: f64, wavenumber: f64, speed: f64) -> Self {
-        let omega = wavenumber * speed; // angular frequency
-        let potential = (wavenumber * r - omega * t).cos() / r;
-
-        // represent as a geometric number with scalar (grade 0) convention
-        Self::from_polar(potential.abs(), if potential >= 0.0 { 0.0 } else { PI })
-    }
     /// computes the derivative of this geometric number with respect to its parameter
     /// using the differential geometric calculus approach
     ///
@@ -806,56 +649,6 @@ impl Geonum {
             normalized_diff - TWO_PI
         } else {
             normalized_diff
-        }
-    }
-
-    /// propagates this geometric number through space and time
-    /// using wave equation principles
-    ///
-    /// this is particularly useful for electromagnetic wave propagation
-    /// where the phase velocity is the speed of light
-    ///
-    /// # arguments
-    /// * `time` - the time coordinate
-    /// * `position` - the spatial coordinate
-    /// * `velocity` - the propagation velocity (usually speed of light)
-    ///
-    /// # returns
-    /// a new geometric number representing the propagated wave
-    pub fn propagate(&self, time: f64, position: f64, velocity: f64) -> Self {
-        // compute phase based on position and time
-        let phase = position - velocity * time;
-
-        // create new geometric number with same length but adjusted angle
-        Geonum {
-            length: self.length,
-            angle: self.angle + phase, // phase modulation
-            blade: self.blade,         // preserve blade grade
-        }
-    }
-
-    /// creates a dispersive wave at a given position and time
-    ///
-    /// computes the phase based on a dispersion relation with wavenumber and frequency
-    /// particularly useful for electromagnetic waves and quantum mechanical waves
-    ///
-    /// # arguments
-    /// * `position` - the spatial coordinate
-    /// * `time` - the time coordinate
-    /// * `wavenumber` - the wavenumber (k) of the wave
-    /// * `frequency` - the angular frequency (ω) of the wave
-    ///
-    /// # returns
-    /// a new geometric number representing the wave at the specified position and time
-    pub fn disperse(position: f64, time: f64, wavenumber: f64, frequency: f64) -> Self {
-        // compute phase based on dispersion relation: φ = kx - ωt
-        let phase = wavenumber * position - frequency * time;
-
-        // create new geometric number with unit length and phase angle
-        Geonum {
-            length: 1.0,
-            angle: phase,
-            blade: 1, // default to vector grade (1)
         }
     }
 
@@ -1324,67 +1117,6 @@ mod tests {
     }
 
     #[test]
-    fn it_propagates() {
-        // create a geometric number representing a wave
-        let wave = Geonum {
-            length: 1.0,
-            angle: 0.0,
-            blade: 1, // vector (grade 1)
-        };
-
-        // define wave parameters
-        let velocity = 3.0e8; // speed of light
-        let time_1 = 0.0;
-        let time_2 = 1.0e-9; // 1 nanosecond later
-        let position = 0.0;
-
-        // propagate wave at two different time points
-        let wave_t1 = wave.propagate(time_1, position, velocity);
-        let wave_t2 = wave.propagate(time_2, position, velocity);
-
-        // verify length is preserved during propagation
-        assert_eq!(
-            wave_t1.length, wave.length,
-            "propagation should preserve amplitude"
-        );
-        assert_eq!(
-            wave_t2.length, wave.length,
-            "propagation should preserve amplitude"
-        );
-
-        // verify phase evolves as expected: phase = position - velocity * time
-        let expected_phase_t1 = position - velocity * time_1;
-        let expected_phase_t2 = position - velocity * time_2;
-
-        assert!(
-            (wave_t1.angle - (wave.angle + expected_phase_t1)).abs() < 1e-10,
-            "phase at t1 should evolve according to position - velocity * time"
-        );
-        assert!(
-            (wave_t2.angle - (wave.angle + expected_phase_t2)).abs() < 1e-10,
-            "phase at t2 should evolve according to position - velocity * time"
-        );
-
-        // verify phase difference between two time points
-        let phase_diff = wave_t2.angle - wave_t1.angle;
-        let expected_diff = -velocity * (time_2 - time_1);
-        assert!(
-            (phase_diff - expected_diff).abs() < 1e-10,
-            "phase difference should equal negative velocity times time difference"
-        );
-
-        // verify propagation in space
-        let position_2 = 1.0; // 1 meter away
-        let wave_p2 = wave.propagate(time_1, position_2, velocity);
-
-        let expected_phase_p2 = position_2 - velocity * time_1;
-        assert!(
-            (wave_p2.angle - (wave.angle + expected_phase_p2)).abs() < 1e-10,
-            "phase at p2 should evolve according to position - velocity * time"
-        );
-    }
-
-    #[test]
     fn it_computes_length_difference() {
         // test length differences between various vectors
         let a = Geonum {
@@ -1596,139 +1328,6 @@ mod tests {
             a.is_orthogonal(&almost),
             "nearly perpendicular vectors should be considered orthogonal"
         );
-    }
-
-    #[test]
-    fn it_disperses() {
-        // define wave parameters
-        let wavenumber = 2.0 * PI; // 2π rad/m (wavelength = 1m)
-        let frequency = 3.0e8 * wavenumber; // ω = c·k for light
-        let position_1 = 0.0;
-        let position_2 = 0.5; // half a wavelength
-        let time_1 = 0.0;
-        let time_2 = 1.0 / (frequency / (2.0 * PI)); // one period
-
-        // create waves at different positions and times
-        let wave_x1_t1 = Geonum::disperse(position_1, time_1, wavenumber, frequency);
-        let wave_x2_t1 = Geonum::disperse(position_2, time_1, wavenumber, frequency);
-        let wave_x1_t2 = Geonum::disperse(position_1, time_2, wavenumber, frequency);
-
-        // verify all waves have unit amplitude
-        assert_eq!(
-            wave_x1_t1.length, 1.0,
-            "dispersed waves should have unit amplitude"
-        );
-
-        // verify phase at origin and t=0 is zero
-        assert!(
-            (wave_x1_t1.angle % TWO_PI).abs() < 1e-10,
-            "phase at origin and t=0 should be zero"
-        );
-
-        // verify spatial phase difference after half a wavelength
-        // phase = kx - ωt, so at t=0, phase difference should be k·(x2-x1) = k·0.5 = π
-        let expected_phase_diff_space = wavenumber * (position_2 - position_1);
-        assert!(
-            (wave_x2_t1.angle - wave_x1_t1.angle - expected_phase_diff_space).abs() < 1e-10,
-            "spatial phase difference should equal wavenumber times distance"
-        );
-
-        // verify temporal phase difference after one period
-        // after one period, the phase should be the same (2π difference)
-        let phase_diff_time = (wave_x1_t2.angle - wave_x1_t1.angle) % TWO_PI;
-        assert!(
-            (phase_diff_time).abs() < 1e-10,
-            "temporal phase should repeat after one period"
-        );
-
-        // verify dispersion relation by comparing wave phase velocities
-        // For k=2π, ω=2πc, wave speed should be c
-        let wave_speed = frequency / wavenumber;
-        let expected_speed = 3.0e8; // speed of light
-        assert!(
-            (wave_speed - expected_speed).abs() / expected_speed < 1e-10,
-            "dispersion relation should yield correct wave speed"
-        );
-    }
-
-    #[test]
-    fn it_computes_electric_field() {
-        // test positive charge
-        let e_field = Geonum::electric_field(2.0, 3.0);
-
-        // coulomb constant
-        let k = 1.0 / (4.0 * PI * VACUUM_PERMITTIVITY);
-
-        // verify magnitude follows inverse square law
-        assert_eq!(e_field.length, k * 2.0 / (3.0 * 3.0));
-
-        // verify direction is outward for positive charge
-        assert_eq!(e_field.angle, PI);
-
-        // test negative charge
-        let e_field_neg = Geonum::electric_field(-2.0, 3.0);
-
-        // verify magnitude is the same
-        assert_eq!(e_field_neg.length, k * 2.0 / (3.0 * 3.0));
-
-        // verify direction is inward for negative charge
-        assert_eq!(e_field_neg.angle, 0.0);
-    }
-
-    #[test]
-    fn it_computes_poynting_vector_with_wedge() {
-        // create perpendicular fields
-        let e = Geonum {
-            length: 5.0,
-            angle: 0.0,
-            blade: 1, // vector (grade 1)
-        }; // along x-axis
-        let b = Geonum {
-            length: 2.0,
-            angle: PI / 2.0,
-            blade: 2, // bivector (grade 2) - magnetic field is a bivector in geometric algebra
-        }; // along y-axis
-
-        let s = e.poynting_vector(&b);
-
-        // check direction is perpendicular to both fields
-        assert_eq!(s.angle, PI); // Using actual wedge product output
-
-        // check magnitude is E×B/μ₀
-        assert_eq!(s.length, (5.0 * 2.0) / VACUUM_PERMEABILITY);
-    }
-
-    #[test]
-    fn it_creates_fields_from_polar_coordinates() {
-        let field = Geonum::from_polar(3.0, PI / 4.0);
-
-        assert_eq!(field.length, 3.0);
-        assert_eq!(field.angle, PI / 4.0);
-
-        let (x, y) = field.to_cartesian();
-        let expected_x = 3.0 * (PI / 4.0).cos();
-        let expected_y = 3.0 * (PI / 4.0).sin();
-
-        assert!((x - expected_x).abs() < 1e-10);
-        assert!((y - expected_y).abs() < 1e-10);
-    }
-
-    #[test]
-    fn it_creates_fields_with_inverse_power_laws() {
-        // test electric field (inverse square)
-        let e_field = Geonum::inverse_field(1.0, 2.0, 2.0, PI, 1.0);
-        assert_eq!(e_field.length, 0.25); // 1.0 * 1.0 / 2.0²
-        assert_eq!(e_field.angle, PI);
-
-        // test gravity (also inverse square)
-        let g_field = Geonum::inverse_field(5.0, 2.0, 2.0, 0.0, 6.67e-11);
-        assert_eq!(g_field.length, 6.67e-11 * 5.0 / 4.0);
-        assert_eq!(g_field.angle, 0.0);
-
-        // test inverse cube field
-        let field = Geonum::inverse_field(2.0, 2.0, 3.0, PI / 2.0, 1.0);
-        assert_eq!(field.length, 0.25); // 1.0 * 2.0 / 2.0³
-        assert_eq!(field.angle, PI / 2.0);
     }
 
     #[test]
