+++ src/lib.rs
@@ -10,6 +10,8 @@
 //! - `projection` - view transformations and projections
 //! - `manifold` - manifold operations and transformations
 //! - `ml` - machine learning operations and neural network functions
+//! - `em` - electromagnetic field calculations and operations
+//! - `waves` - wave propagation and dispersion modeling
 //!
 //! no features are enabled by default, only core functionality
 
@@ -21,15 +23,19 @@ pub mod traits;
 
 // re-export all primary types
 pub use dimensions::Dimensions;
-pub use geonum_mod::{Geonum, EPSILON, TWO_PI, VACUUM_IMPEDANCE};
+pub use geonum_mod::{Geonum, EPSILON, TWO_PI};
 pub use multivector::{Grade, Multivector};
 
 // re-export all traits based on features
+#[cfg(feature = "em")]
+pub use traits::Electromagnetics;
 #[cfg(feature = "manifold")]
 pub use traits::Manifold;
 #[cfg(feature = "optics")]
 pub use traits::Optics;
 #[cfg(feature = "projection")]
 pub use traits::Projection;
+#[cfg(feature = "waves")]
+pub use traits::Waves;
 #[cfg(feature = "ml")]
 pub use traits::{Activation, MachineLearning};
