+++ src/traits/mod.rs
@@ -23,3 +23,13 @@ pub use manifold::Manifold;
 pub mod machine_learning;
 #[cfg(feature = "ml")]
 pub use machine_learning::{Activation, MachineLearning};
+
+#[cfg(feature = "em")]
+pub mod electromagnetics;
+#[cfg(feature = "em")]
+pub use electromagnetics::Electromagnetics;
+
+#[cfg(feature = "waves")]
+pub mod waves;
+#[cfg(feature = "waves")]
+pub use waves::Waves;
